DBAdmin Module Deleted
After much research, I have found that, for some reason, in a NetBeans Platform application, the database Persistence Unit and Entity classes must be compiled to binary from a J2SE project, then included as a wrapped Library Module. Therefore, I have deleted the DBAdmin module, as this was what I was attempting to use for the PU and Entities. I will be creating a Library Module as a wrapper around a compiled J2SE project that will only contain the PU and Entity classes for database access, then I will push that new module up to the repo.

I would love to know why it must be done this way, but cannot seem to find an answer to this question other than, "that is just how it is done." Perhaps a JIRA should be opened with the Apache NetBeans Project to possibly change how this is done so that PUs and Entity classes can just be created within a module. However, I am sure that there may be a valid reason why this is the way it is...

-SC