General Options Started
This commit is just to bring what I have done so far to the GitHub repo. I am switching gears to get the database library done properly so that it will be usable. Once I have the DB Library module pushed, I will go back to work on this module.

-SC