Redone DB Library
OK, so I have recreated the DB Library in accordance with the research I have been doing regarding the problems that I was having. The DB Library Module is now a wrapped library module of the compiled PU and Entity classes that I had created. From all of my research, this should now work, but it is currently untested. Therefore, I will create the DB Admin module (again) to have some viewers and editors for the administrative tables `AccountTypes`, `Accounts` (Chart of Accounts), and `MaintenanceTypes`. By creating these viewers and editors, I will be able to make sure that the DB Library is functioning properly.

-SC