General Options Module
I have created the General Options module, providing a Main Panel (Northwind Traders), as well as a tab on that panel (Company Information). The Company Information tab contains all functionality for loading, storing and validating the settings. For the purposes of Northwind Traders Basic Edition, nothing on this panel is required, unless the user selects the "Have Own Authority" checkbox. Once the user selects that checkbox, all company information is required to be provided.

For the data validation, I have set the background color of invalid fields to a light pinkish-red color as an indication that something is not right. Also, I am providing tip text on the tab, and that text turns red in the event of invalid data. The background color of empty fields, the suite field at all times, when the checkbox is not checked, to a light gold/yellow color. The reason that I am doing this is because, while the data may not be required unless a user has their own authority, it is a good idea to provide the data anyway.

I have also added two (2) validator classes to the `Northwind Utilities` module. These validators are:

* `EmailValidator`: Validates the format of a provided email address, using regular expressions. This class does not attempt to make sure the email address is valid on the internet, just that the form and format of the address is correct.
* `PostalCodeValidator`: Validates US Zip and Canadian Postal Codes. As Northwind Traders expands around the globe (hopeful thinking), we can add more Postal Code formats for different regions into this one `PostalCodeValidator` class. As it stands at this time, the user calls `PostalCodeValidator.isValid(postalCodeField.getText())`. What that call includes is:

```java
public static boolean isValid(String postalCode) {
    return isUsZipCodeValid(postalCode) 
            || isCanadianPostalCodeValid(postalCode);
}
```

Because of the way this method returns `true` or `false`, it would be a simple thing to create and call more Postal Code validation routines in the `return` statement. As long as the Postal Code is valid for one of the validation methods, the `isValid()` method will return `true`. The only way to have `false` returned is if the provided Postal Code is not valid for any of the supported Postal Code regions (currently, US and Canada only).

-SC