Northwind DAO Created
I have created the `NortwindDAO` module for the database system. Per our conversation on Facebook Messenger, I am only creating the database tables that are common to the entire application within this module: Customer/Brokers. I also made a change regarding these two tables, in that a customer and a broker share the vast majority of data, so I just created the one table `CUSTOMERS` and gave it a flag field `BROKER BOOLEAN NOT NULL DEFAULT FALSE`. This way, we only have that one table to maintain and whenever a broker is entered, the broker flag can be set to `true`. Then, instead of executing an SQL statement, such as `SELECT * FROM APP.CUSTOMERS WHERE BROKER = TRUE;`, whenever a list of brokers is needed, I have provided a method called `public List<Customer> getBrokers()` to retrieve only the brokers that are in the table, without any regular customers being included. Likewise, I have created a method called `public List<Customer> getCustomers()` to retrieve all regular customers from the table without any brokers being included. This way, we only need the one table to store shipping/receiving customers, as well as brokers/agents.

The classes exposed by the `NorthwindDAO` module are:

* `com.pekinsoft.nwind.db.api.dao.NorthwindDAO`: provides the mechanisms for creating a connection to the database, getting that connection, creating `Statement` and `PreparedStatement` objects, as well as table creation (with an autogenerated numerical primary key, and with a custom primary key field)(See example below).

* `com.pekinsoft.nwind.db.api.model.Customer`: provides the data model for the `CUSTOMERS` table.

* `com.pekinsoft.nwind.db.api.spi.Customers`: provides the data access object (DAO) for the `CUSTOMERS` table, including:

    * `findAll()`: executes `SELECT * FROM APP.CUSTOMERS`

    * `findBy(String fieldName, T[] value, ComparisonMethod method)`: allows for custom searches of the table. The `T` in this method declaration is only a place-holder for the data type. There are eight (8) unique instances of this method for the following data types:

        * `java.lang.String`

        * `java.lang.Integer`

        * `java.lang.Long`

        * `java.lang.Double`

        * `java.lang.Float`

        * `java.sql.Date`

        * `java.sql.Time`

        * `java.sql.Timestamp`

    * `first()`: moves the record pointer to the first record and returns a `Customer` object

    * `previous()`: moves the record pointer to the previous record and returns a `Customer` object

    * `next()`: moves the record pointer to the next record and returns a `Customer` object

    * `last()`: moves the record pointer to the last record and returns a `Customer` object

    * `update(Customer c)`: updates the current customer record with the data from the supplied `Customer` object

    * `add(Customer c)`: adds a new customer record to the table with the data from the supplied `Customer` object

    * `delete(Customer c)`: deletes the give customer record from the table based upon `results.getInt("ID") == c.getID()`

* `com.pekinsoft.nwind.db.api.enums.ComparisonMethods`: provides an enumeration of valid SQL comparison methods. The allowed methods are:

    * `EQUALS`

    * `NOT_EQUALS`

    * `LESS_THAN`

    * `LESS_THAN_EQUAL_TO`

    * `GREATER_THAN`

    * `GREATER_THAN_EQUAL_TO`

    * `BETWEEN`

    * `LIKE`

The two (2) ways of creating tables are:

1. with an autogenerated primary key (Example 1)

2. with a custom primary key (Example 2)

#### Example 1
```java
NorthwindDAO dao = new NorthwindDAO();

// Create the table allowing for a default primary key field that is an autogenerated number:
dao.createTable("CUSTOMERS", "COMPANY VARCHAR(30) NOT NULL, ADDRESS NOT NULL, CONTACT, PHONE, EMAIL, NOTES)");
///

#### Example 2
```java
NorthwindDAO dao = new NorthwindDAO();

// Create the table allowing for a custom primary key field, which does not need to be numerical:
dao.createTable("CREATE TABLE APP.LOADS(TRIP_NUMBER VARCHAR(15) NOT NULL UNIQUE PRIMARY KEY, COMMODITY, MILES INTEGER NOT NULL, RATE DOUBLE NOT NULL)");
```

These two different ways of creating the tables should only be done once the module has checked to see if the table exists. If it does, then the table does not need to be created. To make sure that the users' data is protected against deletion from over-eager table creation, I have not made the creation of the tables start by dropping the table. This way, if a module does not first check for the existence of the table, the call to either `createTable` method will fail with an `SQLException`. That `SQLException` will be displayed by the `Exceptions.printStacktrace` method, so that we should find our failure during regular testing, before the application comes to a release.

Each of the "functional" modules (i.e., Loads, FuelMgmt, ExpenseMgmt, etc.) will add their own tables to the database. However, the biggest module for adding tables to the database will be the `Accounting` module. The `Accounting` module will add three (3) tables to the database:

* `ACCOUNT_TYPES`: the types of accounts, i.e., expense, income, asset, liability

* `ACCOUNTS`: the chart of accounts for the Basic Edition

* `GEN_LEDG`: the General Ledger, which is the heart of the accounting system

All of the "functional" modules will provide their own journal table for the database. In this way, if we modify a "functional" module's data, it will handle its own database table, as well as its model and controller objects. Therefore, we will only need to update that one module to make the changes effective, without the need of updating two or three modules to make the change. I believe this will ease our maintenance of the system as a whole. 

##### Caveat
The only issue I can see with handling data creation in this manner is if the modules do not load in a particular order. I need to look into whether or not the order the modules are loaded can be controlled in some way. For example, we would not want the `LoadMgmt` module to load prior to the `NorthwindDAO` module. Why? Because, if the `LoadMgmt` module loads first and attempts to create its tables in the database, but the `NorthwindDAO` module has not loaded, then there will be no database (or database connection) in which to create the `LoadMgmt` tables. Therefore, it may be that I will need to modify the `NorthwindDAO` module in the future to simply provide all of the database tables from it upon first run of the application. Then the question will become whether to include the model classes from the `NorthwindDAO` module, or from the "functional" module that supports that table.

With all of the above in mind, let's discuss this on this PR. We do need to come to a decision on the best way forward.

-SC